Feat/61 fixing audit (#64)
* feat(SystemRole): transfering the ownership modifier inside the systemRole contract

* feat(SystemRole): removing dangerous renounceOwnerhsip function

* fix(StandardController): Adding function to remove a bridge frontend

* fix(SafeMath): removing SafeMath from the codebase

* fix(uint): use uint256 where uint was written

* fix(pragma): fixing the pragma version to 8.11

* feat(*Controller): adding underscore for internal function

* feat(Claimable): Adding new event for Pending state variable change

* feat(SystemRole/MintableController): tranfering mintAllowance feature from systemRole to mintableController

* feat(MintableController/SmartTokenLib): checking return value for token function

* feat(ClaimableEvent): Renaming events

+++ contracts/IValidator.sol
@@ -32,7 +32,7 @@ interface IValidator {
     event Decision(
         address indexed from,
         address indexed to,
-        uint amount,
+        uint256 amount,
         bool valid
     );
 
@@ -46,6 +46,6 @@ interface IValidator {
     function validate(
         address from,
         address to,
-        uint amount
+        uint256 amount
     ) external returns (bool valid);
 }
